ecom_backend/requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
numpy==1.21.4
pandas==1.3.5
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
